Update 05-TrustLNGateways.md
It seems like the intended message was that if an LN gateway was unresponsive or malicious, a user could choose an alternate LN gateway. However, in the document it says that if the "mint" is down a user could utilize an alternative LN gateway. Unless I'm misunderstanding this mechanism, the wording seems incorrect and should be updated.